refactor: 예약 생성 책임 분리
