+++ src/main/java/roomescape/service/ReservationService.java
@@ -28,11 +28,13 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
     private final ThemeRepository themeRepository;
+    private final ReservationCreator reservationCreator;
 
-    public ReservationService(ReservationRepository reservationRepository, ReservationTimeRepository reservationTimeRepository, ThemeRepository themeRepository) {
+    public ReservationService(ReservationRepository reservationRepository, ReservationTimeRepository reservationTimeRepository, ThemeRepository themeRepository, ReservationCreator reservationCreator) {
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.themeRepository = themeRepository;
+        this.reservationCreator = reservationCreator;
     }
 
     public List<Reservation> findByName(String name) {
@@ -41,14 +43,9 @@ public class ReservationService {
 
     @Transactional
     public Reservation create(String name, LocalDate date, Long timeId, Long themeId) {
-        ReservationTime time = findReservationTime(timeId);
+        ReservationTime time = reservationCreator.findReservationTime(timeId);
         validateNotPast(date, time);
-        validateAlreadyReserved(date, timeId, themeId);
-        Theme theme = findTheme(themeId);
-        Reservation reservation = new Reservation(null, name, date, time, theme);
-        Long id = reservationRepository.insert(reservation);
-        return reservationRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("생성된 예약을 찾을 수 없습니다."));
+        return reservationCreator.create(name, date, timeId, themeId);
     }
 
     @Transactional
@@ -85,11 +82,6 @@ public class ReservationService {
                 .toList();
     }
 
-    private ReservationTime findReservationTime(Long timeId) {
-        return reservationTimeRepository.findBy(timeId)
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 예약 시간입니다."));
-    }
-
     private Reservation findReservation(Long id) {
         return reservationRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("존재하지 않는 예약입니다."));
@@ -122,7 +114,7 @@ public class ReservationService {
                 .noneMatch(reservation -> reservation.hasTime(time));
     }
 
-    private void validateAlreadyReserved(LocalDate date, Long timeId, Long themeId) {
+    private void validateUpdatableReservationAlreadyReserved(LocalDate date, Long timeId, Long themeId) {
         if (reservationRepository.existsWith(date, timeId, themeId)) {
             throw new DuplicateReservationException("이미 예약된 시간입니다.");
         }
@@ -136,7 +128,7 @@ public class ReservationService {
     private void validateUpdatePolicy(Reservation reservation, Reservation updatedReservation) {
         validateScheduleChanged(reservation, updatedReservation);
         validateNotPast(updatedReservation.getDate(), updatedReservation.getTime());
-        validateAlreadyReserved(
+        validateUpdatableReservationAlreadyReserved(
                 updatedReservation.getDate(),
                 updatedReservation.getTime().getId(),
                 updatedReservation.getTheme().getId());
@@ -174,7 +166,7 @@ public class ReservationService {
 
     private ReservationTime resolveUpdateTime(Reservation reservation, Long timeId) {
         if (timeId != null) {
-            return findReservationTime(timeId);
+            return reservationCreator.findReservationTime(timeId);
         }
         return reservation.getTime();
     }
